refactor(astutil): AST 周りの共通処理を util へ移動
