+++ src/feature/metric/cognitiveComplexity/CognitiveComplexity.spec.ts
@@ -1,7 +1,7 @@
 import { test, expect } from 'vitest';
-import AstLogger from '../AstLogger';
+import AstLogger from '../../util/AstLogger';
 import * as ts from 'typescript';
-import AstTraverser from '../AstTraverser';
+import AstTraverser from '../../util/AstTraverser';
 import { readFileSync } from 'fs';
 import CognitiveComplexityForSourceCode from './CognitiveComplexityForSourceCode';
 import { createCognitiveComplexityAnalyzer } from '../cognitiveComplexity';
