+++ src/feature/metric/cyclomaticComplexity/CyclomaticComplexity.spec.ts
@@ -1,7 +1,7 @@
 import { test, expect } from 'vitest';
-import AstLogger from '../AstLogger';
+import AstLogger from '../../util/AstLogger';
 import * as ts from 'typescript';
-import AstTraverser from '../AstTraverser';
+import AstTraverser from '../../util/AstTraverser';
 import { createCyclomaticComplexityAnalyzer } from '.';
 
 type OperatorTest = {
